swab32(): Fix a very serious bug.
git-svn-id: https://src.develer.com/svnoss/bertos/trunk@77 38d2e660-2303-0410-9eaa-f027e97ec537

+++ mware/byteorder.h
@@ -15,6 +15,9 @@
 
 /*
  * $Log$
+ * Revision 1.4  2004/07/22 01:08:43  bernie
+ * swab32(): Fix a very serious bug.
+ *
  * Revision 1.3  2004/07/20 23:47:12  bernie
  * Finally remove redundant protos.
  *
@@ -44,7 +47,7 @@ INLINE uint16_t swab16(uint16_t x)
 /*!
  * \brief Reverse bytes in a 32-bit value (e.g.: 0x12345678 -> 0x78563412).
  */
-INLINE uint16_t swab32(uint16_t x)
+INLINE uint32_t swab32(uint32_t x)
 {
 	return    ((x & (uint32_t)0x000000FFUL) << 24)
 		| ((x & (uint32_t)0x0000FF00UL) <<  8)
